Use size_t over (unsigned) long for container size
Amongst other things, switch mod_manager_ui selection from long to
size_t

This was written as if it might be negative, but none of the callers
could ever make it so, and the code is just much simpler if it's a
size_t.

+++ src/monster.cpp
@@ -599,7 +599,7 @@ int monster::print_info( const catacurses::window &w, int vStart, int vLines, in
     }
 
     std::string effects = get_effect_status();
-    long long used_space = att.first.length() + name().length() + 3;
+    size_t used_space = att.first.length() + name().length() + 3;
     trim_and_print( w, vStart++, used_space, getmaxx( w ) - used_space - 2,
                     h_white, effects );
